fix data types

+++ get_next_line_bonus.h
@@ -6,7 +6,7 @@
 /*   By: stakada <stakada@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/01 11:38:35 by stakada           #+#    #+#             */
-/*   Updated: 2024/06/01 12:16:56 by stakada          ###   ########.fr       */
+/*   Updated: 2024/06/01 14:37:59 by stakada          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,12 +26,12 @@
 
 // TODO 提出前に更新
 char	*get_next_line(int fd);
-int		find_nl(char *str);
+ssize_t	find_nl(char *str);
 char	*join_read(char *s1, char *s2);
 char	*divide_string(char **store);
 size_t	ft_strlen_gnl(const char *s);
 void	ft_strcpy_gnl(char *dst, char *src);
-void	ft_strncpy_gnl(char *dst, char *src, int n);
+void	ft_strncpy_gnl(char *dst, char *src, size_t n);
 char	*ft_strdup_gnl(char *src);
 
 #endif
